requirements.txt
youtube-transcript-api==0.6.2
zipp==3.19.2
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.55.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.6.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
streamlit==1.35.0
streamlit-drawable-canvas==0.9.3
streamlit-webrtc==0.47.7
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1